requirements.txt
streamlit>=1.28.0
moviepy>=1.0.3
openai-whisper>=20230314
pysrt>=1.1.2
torch>=2.0.0
numpy>=1.24.0
Pillow>=9.5.0
imageio>=2.25.0
imageio-ffmpeg>=0.4.9
scipy>=1.10.0
